[SPARK-44691][SQL][CONNECT] Move Subclasses of Analysis to sql/api
