+++ sql/api/src/main/scala/org/apache/spark/sql/catalyst/util/AttributeNameParser.scala
@@ -16,8 +16,6 @@
  */
 package org.apache.spark.sql.catalyst.util
 
-import org.apache.spark.sql.errors.DataTypeErrors
-
 trait AttributeNameParser {
   /**
    * Used to split attribute name by dot with backticks rule.
@@ -26,42 +24,7 @@ trait AttributeNameParser {
    * We can use backtick only inside quoted name parts.
    */
   def parseAttributeName(name: String): Seq[String] = {
-    def e = DataTypeErrors.attributeNameSyntaxError(name)
-    val nameParts = scala.collection.mutable.ArrayBuffer.empty[String]
-    val tmp = scala.collection.mutable.ArrayBuffer.empty[Char]
-    var inBacktick = false
-    var i = 0
-    while (i < name.length) {
-      val char = name(i)
-      if (inBacktick) {
-        if (char == '`') {
-          if (i + 1 < name.length && name(i + 1) == '`') {
-            tmp += '`'
-            i += 1
-          } else {
-            inBacktick = false
-            if (i + 1 < name.length && name(i + 1) != '.') throw e
-          }
-        } else {
-          tmp += char
-        }
-      } else {
-        if (char == '`') {
-          if (tmp.nonEmpty) throw e
-          inBacktick = true
-        } else if (char == '.') {
-          if (name(i - 1) == '.' || i == name.length - 1) throw e
-          nameParts += tmp.mkString
-          tmp.clear()
-        } else {
-          tmp += char
-        }
-      }
-      i += 1
-    }
-    if (inBacktick) throw e
-    nameParts += tmp.mkString
-    nameParts.toSeq
+    SparkStringUtils.parseAttributeName(name)
   }
 }
