+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/AlreadyExistException.scala
@@ -22,120 +22,7 @@ import org.apache.spark.sql.AnalysisException
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.catalog.CatalogTypes.TablePartitionSpec
 import org.apache.spark.sql.catalyst.util.{quoteIdentifier, quoteNameParts }
-import org.apache.spark.sql.connector.catalog.CatalogV2Implicits.IdentifierHelper
-import org.apache.spark.sql.connector.catalog.Identifier
 import org.apache.spark.sql.types.StructType
-
-/**
- * Thrown by a catalog when an item already exists. The analyzer will rethrow the exception
- * as an [[org.apache.spark.sql.AnalysisException]] with the correct position information.
- */
-class DatabaseAlreadyExistsException(db: String)
-  extends NamespaceAlreadyExistsException(Array(db))
-
-// any changes to this class should be backward compatible as it may be used by external connectors
-class NamespaceAlreadyExistsException private(
-    message: String,
-    errorClass: Option[String],
-    messageParameters: Map[String, String])
-  extends AnalysisException(
-    message,
-    errorClass = errorClass,
-    messageParameters = messageParameters) {
-
-  def this(errorClass: String, messageParameters: Map[String, String]) = {
-    this(
-      SparkThrowableHelper.getMessage(errorClass, messageParameters),
-      Some(errorClass),
-      messageParameters)
-  }
-
-  def this(namespace: Array[String]) = {
-    this(errorClass = "SCHEMA_ALREADY_EXISTS",
-      Map("schemaName" -> quoteNameParts(namespace)))
-  }
-
-  def this(message: String) = {
-    this(
-      message,
-      errorClass = Some("SCHEMA_ALREADY_EXISTS"),
-      messageParameters = Map.empty[String, String])
-  }
-}
-
-// any changes to this class should be backward compatible as it may be used by external connectors
-class TableAlreadyExistsException private(
-    message: String,
-    cause: Option[Throwable],
-    errorClass: Option[String],
-    messageParameters: Map[String, String])
-  extends AnalysisException(
-    message,
-    cause = cause,
-    errorClass = errorClass,
-    messageParameters = messageParameters) {
-
-  def this(errorClass: String, messageParameters: Map[String, String], cause: Option[Throwable]) = {
-    this(
-      SparkThrowableHelper.getMessage(errorClass, messageParameters),
-      cause,
-      Some(errorClass),
-      messageParameters)
-  }
-
-  def this(db: String, table: String) = {
-    this(errorClass = "TABLE_OR_VIEW_ALREADY_EXISTS",
-      messageParameters = Map("relationName" ->
-        (quoteIdentifier(db) + "." + quoteIdentifier(table))),
-      cause = None)
-  }
-
-  def this(table: String) = {
-    this(errorClass = "TABLE_OR_VIEW_ALREADY_EXISTS",
-      messageParameters = Map("relationName" ->
-        quoteNameParts(UnresolvedAttribute.parseAttributeName(table))),
-      cause = None)
-  }
-
-  def this(table: Seq[String]) = {
-    this(errorClass = "TABLE_OR_VIEW_ALREADY_EXISTS",
-      messageParameters = Map("relationName" -> quoteNameParts(table)),
-      cause = None)
-  }
-
-  def this(tableIdent: Identifier) = {
-    this(errorClass = "TABLE_OR_VIEW_ALREADY_EXISTS",
-      messageParameters = Map("relationName" -> tableIdent.quoted),
-      cause = None)
-  }
-
-  def this(message: String, cause: Option[Throwable] = None) = {
-    this(
-      message,
-      cause,
-      errorClass = Some("TABLE_OR_VIEW_ALREADY_EXISTS"),
-      messageParameters = Map.empty[String, String])
-  }
-}
-
-class TempTableAlreadyExistsException(errorClass: String, messageParameters: Map[String, String],
-                                      cause: Option[Throwable] = None)
-  extends AnalysisException(errorClass, messageParameters, cause = cause) {
-  def this(table: String) = {
-    this(errorClass = "TEMP_TABLE_OR_VIEW_ALREADY_EXISTS",
-      messageParameters = Map("relationName"
-        -> quoteNameParts(UnresolvedAttribute.parseAttributeName(table))))
-  }
-}
-
-// any changes to this class should be backward compatible as it may be used by external connectors
-class ViewAlreadyExistsException(errorClass: String, messageParameters: Map[String, String])
-  extends AnalysisException(errorClass, messageParameters) {
-
-  def this(ident: Identifier) =
-    this(errorClass = "VIEW_ALREADY_EXISTS",
-      messageParameters = Map("relationName" -> ident.quoted))
-}
 
 // any changes to this class should be backward compatible as it may be used by external connectors
 class PartitionAlreadyExistsException private(
@@ -223,56 +110,6 @@ class PartitionsAlreadyExistException private(
     this(
       message,
       errorClass = Some("PARTITIONS_ALREADY_EXIST"),
-      messageParameters = Map.empty[String, String])
-  }
-}
-
-// any changes to this class should be backward compatible as it may be used by external connectors
-class FunctionAlreadyExistsException(errorClass: String, messageParameters: Map[String, String])
-  extends AnalysisException(errorClass, messageParameters) {
-
-  def this(function: Seq[String]) = {
-    this (errorClass = "ROUTINE_ALREADY_EXISTS",
-      Map("routineName" -> quoteNameParts(function)))
-  }
-
-  def this(db: String, func: String) = {
-    this(Seq(db, func))
-  }
-}
-
-// any changes to this class should be backward compatible as it may be used by external connectors
-class IndexAlreadyExistsException private(
-    message: String,
-    cause: Option[Throwable],
-    errorClass: Option[String],
-    messageParameters: Map[String, String])
-  extends AnalysisException(
-    message,
-    cause = cause,
-    errorClass = errorClass,
-    messageParameters = messageParameters) {
-
-  def this(
-      errorClass: String,
-      messageParameters: Map[String, String],
-      cause: Option[Throwable]) = {
-    this(
-      SparkThrowableHelper.getMessage(errorClass, messageParameters),
-      cause,
-      Some(errorClass),
-      messageParameters)
-  }
-
-  def this(indexName: String, tableName: String, cause: Option[Throwable]) = {
-    this("INDEX_ALREADY_EXISTS", Map("indexName" -> indexName, "tableName" -> tableName), cause)
-  }
-
-  def this(message: String, cause: Option[Throwable] = None) = {
-    this(
-      message,
-      cause,
-      errorClass = Some("INDEX_ALREADY_EXISTS"),
       messageParameters = Map.empty[String, String])
   }
 }
